+++ cmn-data-svc-client/src/test/java/org/acumos/cds/client/test/DomainTest.java
@@ -716,7 +716,6 @@ public class DomainTest extends AbstractModelTest {
 	public void testMLPSolution() {
 		MLPSolution m = new MLPSolution(s1, s1, b1);
 		m = new MLPSolution();
-		m.setAccessTypeCode(s1);
 		m.setActive(b1);
 		m.setCreated(d1);
 		m.setDescription(s2);
@@ -731,7 +730,6 @@ public class DomainTest extends AbstractModelTest {
 		m.setSolutionId(s10);
 		m.setToolkitTypeCode(s11);
 		m.setValidationStatusCode(s12);
-		Assert.assertEquals(s1, m.getAccessTypeCode());
 		Assert.assertEquals(b1, m.isActive());
 		Assert.assertEquals(d1, m.getCreated());
 		Assert.assertEquals(s2, m.getDescription());
@@ -902,35 +900,37 @@ public class DomainTest extends AbstractModelTest {
 
 	@Test
 	public void testMLPSolutionRevision() {
-		MLPSolutionRevision m = new MLPSolutionRevision(s1, s1, s1);
+		MLPSolutionRevision m = new MLPSolutionRevision(s1, s1, s1, s1);
 		m = new MLPSolutionRevision();
+		m.setAccessTypeCode(s1);
 		m.setCreated(d1);
-		m.setDescription(s1);
-		m.setMetadata(s2);
+		m.setDescription(s2);
+		m.setMetadata(s3);
 		m.setModified(d2);
-		m.setOrigin(s3);
-		m.setOwnerId(s4);
-		m.setRevisionId(s5);
-		m.setSolutionId(s6);
-		m.setSourceId(s7);
-		m.setVersion(s8);
+		m.setOrigin(s4);
+		m.setOwnerId(s5);
+		m.setRevisionId(s6);
+		m.setSolutionId(s7);
+		m.setSourceId(s8);
+		m.setVersion(s9);
 		Assert.assertEquals(d1, m.getCreated());
-		Assert.assertEquals(s1, m.getDescription());
-		Assert.assertEquals(s2, m.getMetadata());
+		Assert.assertEquals(s1, m.getAccessTypeCode());
+		Assert.assertEquals(s2, m.getDescription());
+		Assert.assertEquals(s3, m.getMetadata());
 		Assert.assertEquals(d2, m.getModified());
-		Assert.assertEquals(s3, m.getOrigin());
-		Assert.assertEquals(s4, m.getOwnerId());
-		Assert.assertEquals(s5, m.getRevisionId());
-		Assert.assertEquals(s6, m.getSolutionId());
-		Assert.assertEquals(s7, m.getSourceId());
-		Assert.assertEquals(s8, m.getVersion());
+		Assert.assertEquals(s4, m.getOrigin());
+		Assert.assertEquals(s5, m.getOwnerId());
+		Assert.assertEquals(s6, m.getRevisionId());
+		Assert.assertEquals(s7, m.getSolutionId());
+		Assert.assertEquals(s8, m.getSourceId());
+		Assert.assertEquals(s9, m.getVersion());
 		Assert.assertFalse(m.equals(null));
 		Assert.assertFalse(m.equals(new Object()));
 		Assert.assertTrue(m.equals(m));
 		Assert.assertNotNull(m.hashCode());
 		logger.info(m.toString());
 		try {
-			new MLPSolutionRevision(null, null, null);
+			new MLPSolutionRevision(null, null, null, null);
 			Assert.assertTrue("Unexpected success", false);
 		} catch (IllegalArgumentException iae) {
 			// null arg is rejected